split some document-related code from core.js to src/Document.js
