+++ core.js
@@ -7,7 +7,6 @@ const child_process = require('child_process');
 const os = require('os');
 const util = require('util');
 const {URL} = url;
-const parse5 = require('parse5');
 const {performance} = require('perf_hooks');
 const {XMLHttpRequest: XMLHttpRequestBase, FormData} = require('window-xhr');
 
@@ -42,7 +41,8 @@ const symbols = require('./src/symbols');
 const {urls} = require('./src/urls');
 
 // Class imports.
-const {Document, DocumentFragment} = require('./src/Document');
+const {_parseDocument, _parseDocumentAst, Document, DocumentFragment, DocumentType,
+       DOMImplementation, initDocument} = require('./src/Document');
 const DOM = require('./src/DOM');
 const {DOMRect, Node, NodeList} = require('./src/DOM');
 const {CustomEvent, DragEvent, Event, EventTarget, KeyboardEvent, MessageEvent, MouseEvent,
@@ -143,6 +143,7 @@ class Resources extends EventTarget {
     return resource;
   }
 }
+GlobalContext.Resources = Resources;
 
 class CustomElementRegistry {
   constructor(window) {
@@ -686,29 +687,6 @@ class MediaRecorder extends EventEmitter {
 }
 
 class MicrophoneMediaStream {}
-
-class DocumentType {}
-class DOMImplementation {
-  constructor(window) {
-    this._window = window;
-  }
-
-  createDocument() {
-    throw new Error('not implemented');
-  }
-
-  createDocumentType() {
-    return new DocumentType();
-  }
-
-  createHTMLDocument() {
-    return _parseDocument('', this._window[symbols.optionsSymbol], this._window);
-  }
-
-  hasFeature() {
-    return false;
-  }
-}
 
 class DataTransfer {
   constructor({items = [], files = []} = {}) {
@@ -1800,146 +1778,6 @@ const _makeWindow = (options = {}, parent = null, top = null) => {
   return window;
 };
 GlobalContext._makeWindow = _makeWindow;
-const _parseDocument = GlobalContext._parseDocument = (s, options, window) => {
-  const ast = parse5.parse(s, {
-    sourceCodeLocationInfo: true,
-  });
-  ast.tagName = 'document';
-  return _parseDocumentAst(ast, options, window, true);
-};
-
-const _parseDocumentAst = (ast, options, window, uppercase) => {
-  const document = _fromAST(ast, window, null, null, uppercase);
-  const html = document.childNodes.find(el => el.tagName === 'HTML');
-const documentElement = html || (document.childNodes.length > 0 ? document.childNodes[0] : null);
-  const head = html ? html.childNodes.find(el => el.tagName === 'HEAD') : null;
-  const body = html ? html.childNodes.find(el => el.tagName === 'BODY') : null;
-
-  document.documentElement = documentElement;
-  document.readyState = 'loading';
-  document.head = head;
-  document.body = body;
-  document.location = window.location;
-  document.createElement = tagName => {
-    tagName = tagName.toUpperCase();
-    const HTMLElementTemplate = window[symbols.htmlTagsSymbol][tagName];
-    const element = HTMLElementTemplate ? new HTMLElementTemplate() : new DOM.HTMLElement(tagName);
-    element.ownerDocument = document;
-    return element;
-  };
-  document.createElementNS = (namespace, tagName) => document.createElement(tagName);
-  document.createDocumentFragment = () => {
-    const documentFragment = new DocumentFragment();
-    documentFragment.ownerDocument = document;
-    return documentFragment;
-  };
-  document.createTextNode = text => new DOM.Text(text);
-  document.createComment = comment => new DOM.Comment(comment);
-  document.createEvent = type => {
-    switch (type) {
-      case 'KeyboardEvent':
-      case 'KeyboardEvents':
-        return new KeyboardEvent();
-      case 'MouseEvent':
-      case 'MouseEvents':
-        return new MouseEvent();
-      case 'Event':
-      case 'Events':
-      case 'HTMLEvents':
-        return new Event();
-      default:
-        throw new Error('invalid createEvent type: ' + type);
-    }
-  };
-  document.importNode = (el, deep) => el.cloneNode(deep);
-  document.scripts = utils._makeHtmlCollectionProxy(document.documentElement, 'script');
-  document.styleSheets = [];
-  document.implementation = new DOMImplementation(window);
-  document.resources = new Resources(); // non-standard
-  document.activeElement = body;
-  document.open = () => {
-    document.innerHTML = '';
-  };
-  document.close = () => {};
-  document.write = htmlString => {
-    const childNodes = parse5.parseFragment(htmlString, {
-      locationInfo: true,
-    }).childNodes.map(childNode => _fromAST(childNode, window, document.body, document, true));
-    for (let i = 0; i < childNodes.length; i++) {
-      document.body.appendChild(childNodes[i]);
-    }
-  };
-  document.execCommand = command => {
-    if (command === 'copy') {
-      // nothing
-    } else if (command === 'paste') {
-      document.dispatchEvent(new Event('paste'));
-    }
-  };
-  document[symbols.pointerLockElementSymbol] = null;
-  document[symbols.fullscreenElementSymbol] = null;
-
-  if (window.top === window) {
-    document.addEventListener('pointerlockchange', () => {
-      const iframes = document.getElementsByTagName('iframe');
-      for (let i = 0; i < iframes.length; i++) {
-        const iframe = iframes[i];
-        if (iframe.contentDocument) {
-          iframe.contentDocument._emit('pointerlockchange');
-        }
-      }
-    });
-    document.addEventListener('fullscreenchange', () => {
-      const iframes = document.getElementsByTagName('iframe');
-      for (let i = 0; i < iframes.length; i++) {
-        const iframe = iframes[i];
-        if (iframe.contentDocument) {
-          iframe.contentDocument._emit('fullscreenchange');
-        }
-      }
-    });
-  }
-
-  process.nextTick(async () => {
-    if (body) {
-      const bodyChildNodes = body.childNodes;
-      body.childNodes = new NodeList();
-
-      try {
-        await _runHtml(document.head, window);
-      } catch(err) {
-        console.warn(err);
-      }
-
-      body.childNodes = bodyChildNodes;
-
-      document.dispatchEvent(new Event('DOMContentLoaded', {target: document}));
-
-      try {
-        await _runHtml(document.body, window);
-      } catch(err) {
-        console.warn(err);
-      }
-
-      document.readyState = 'interactive';
-      document.dispatchEvent(new Event('readystatechange', {target: document}));
-
-      document.readyState = 'complete';
-      document.dispatchEvent(new Event('readystatechange', {target: document}));
-
-      document.dispatchEvent(new Event('load', {target: document}));
-      window.dispatchEvent(new Event('load', {target: window}));
-    } else {
-      try {
-        await _runHtml(document, window);
-      } catch(err) {
-        console.warn(err);
-      }
-    }
-  });
-
-  return document;
-};
 
 const _makeWindowWithDocument = (s, options, parent, top) => {
   const window = _makeWindow(options, parent, top);
